prettyping.sh: Added support for non-terminal output.
This is the messiest change I've implemented. It adds lots of "if is terminal"
branches all over the code.  With this feature, when the output is not a
terminal, cursor-positioning escapes will not be used. This means the
statistics bar must be printed from a very different code path (once every full
dotted line, instead of on every packet).

Now it is possible to save the output to a file, or send it through a pipe. It
is even possible to run prettyping.sh from within Gvim (by passing --nocolor
--noterminal).