Started writing a new testing framework for prettyping.
This new framework is written in Python and should be more flexible. It
should support multiple test cases and it should be easy to test them.

However, this will not be automated testing. Instead, it will run
prettyping with specific input (i.e. a mock ping binary), and a human
(me!) will observe if the output and the prettyping behavior makes
sense.