Add support for IPv6 on OS X/BSD
Though prettyping works out of the box with v6 on Linux the BSD `ping6`
utility has slightly different output.

In ICMPv6 what we know as the TTL in IPv6 is known as the hop limit, or
`hlim` in the output. It also uses a comma after the IP instead of a
colon as that can get a bit confusing with IPv6 addresses.