Swapping the order of numbers and characters in the legend.
Previously, the legend looked like this:
X   0 Y  10 Z

Which was counter-intuive. (see #6)

David (iamsudo) suggested putting numbers in the middle:

0 X 10 Y 20 Z

Which means that X is between 0 and 10, which maps precisely to the
prettyping behavior (in this case, 0 <= X < 10).

Hopefully, this will be less confusing. I think this also solves #6.